backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
supabase>=2.11.0
python-dotenv>=1.0.0
pydantic-settings>=2.7.0
cryptography>=44.0.0
PyJWT>=2.9.0
sse-starlette>=2.2.0
